+++ packages/core/src/_test/e2e/factory/factory.test.ts
@@ -3,7 +3,7 @@ import { Ponder } from "@/Ponder.js";
 import { setupAnvil, setupIsolatedDatabase } from "@/_test/setup.js";
 import { simulate } from "@/_test/simulate.js";
 import { onAllEventsIndexed } from "@/_test/utils.js";
-import { buildOptions } from "@/config/options.js";
+import { buildOptions } from "@/common/options.js";
 import { range } from "@/utils/range.js";
 import request from "supertest";
 import { afterEach, beforeEach, expect, test } from "vitest";
