+++ packages/core/src/common/codegen.ts
@@ -1,3 +1,9 @@
+import { mkdirSync, writeFileSync } from "node:fs";
+import path from "node:path";
+import type { Common } from "@/Ponder.js";
+import type { GraphQLSchema } from "graphql";
+import { printSchema } from "graphql";
+
 export const ponderEnv = `// This file enables type checking and editor autocomplete for this Ponder project.
 // After upgrading, you may find that changes have been made to this file.
 // If this happens, please commit the changes. Do not manually edit this file.
@@ -26,3 +32,34 @@ declare module "@/generated" {
   export type Schema = Virtual.Schema<schema>;
 }
 `;
+
+export function codegen({
+  common,
+  graphqlSchema,
+}: {
+  common: Common;
+  graphqlSchema: GraphQLSchema;
+}) {
+  writeFileSync(
+    path.join(common.options.rootDir, "ponder-env.d.ts"),
+    ponderEnv,
+    "utf8",
+  );
+
+  common.logger.debug({
+    service: "codegen",
+    msg: "Wrote new file at ponder-env.d.ts",
+  });
+
+  mkdirSync(common.options.generatedDir);
+  writeFileSync(
+    path.join(common.options.generatedDir, "schema.graphql"),
+    printSchema(graphqlSchema),
+    "utf-8",
+  );
+
+  common.logger.debug({
+    service: "codegen",
+    msg: "Wrote new file at generated/schema.graphql",
+  });
+}
