+++ packages/core/src/common/errors.ts
