+++ packages/core/src/common/options.ts
